app/requirements.txt
fastapi==0.112.2
python-dotenv==1.0.1
uvicorn==0.30.6
pymongo==4.8.0